Enabled more warnings and cleaned them

+++ dasher.c
@@ -9,8 +9,8 @@
 #define BYTE unsigned char
 #define WORD unsigned int
 #define SIZE unsigned long long
-#define SW sizeof(WORD)
-#define SS sizeof(SIZE)
+#define SW (int)sizeof(WORD)
+#define SS (int)sizeof(SIZE)
 #define NI 16
 #if RIPE==0 || RIPE==128
 #define NO 4
@@ -773,7 +773,7 @@ static void next(const BYTE ba[NI*SW])
 	size += NI*SW;
 }
 
-static void last(const BYTE ba[NI*SW],size_t nb)
+static void last(const BYTE ba[NI*SW],int nb)
 {
 	int i,j,k,l;
 	BYTE temp[2*NI*SW];
